pyproject.toml
[project]
name = "stochastic-swarm-viz"
version = "1.0.0"
description = "Visualization tools for Stochastic Swarm particle simulations"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "matplotlib",
    "gymnasium",
    "pybind11",
    "pytest",
    "stable-baselines3",
    "tensorboard",
    "tqdm",
    "rich",
    "torch",
    "pytorch-lightning"
]

[project.optional-dependencies]
dev = [
    "ipython",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["python"]

[tool.uv]
dev-dependencies = []
